# ivr_api/docs/swagger.yaml
openapi: 3.0.0
info:
  title: IVR Calling System API
  description: |
    A comprehensive IVR (Interactive Voice Response) calling system for marketing campaigns with multilanguage support.

    ## Features
    - Campaign management (CRUD operations)
    - Bulk call initiation with Twilio integration
    - Multi-language support (English, Spanish, French, German, Hindi)
    - Real-time call status tracking
    - IVR menu navigation with user input handling
    - Opt-out management

    ## Authentication
    Currently, the API is open. For production, implement JWT or API key authentication.

    ## Rate Limiting
    Not implemented. Recommended for production use.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.yourcompany.com
    description: Production server
...
paths:
...
  /api/campaigns/{id}/calls:
    get:
      tags:
        - Campaigns
      summary: Get campaign calls
      description: Retrieves all calls associated with a specific campaign and statistics
      operationId: getCampaignCalls
      parameters:
        - name: id
          in: path
          required: true
          description: Campaign ID (MongoDB ObjectID)
          schema:
            type: string
            example: 507f1f77bcf86cd799439011
      responses:
        "200":
          description: Campaign calls and statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  calls:
                    type: array
                    items:
                      $ref: "#/components/schemas/Call"
                  statistics:
                    type: object
                    properties:
                      total:
                        type: integer
                        example: 100
                      pending:
                        type: integer
                        example: 10
                      initiated:
                        type: integer
                        example: 30
                      completed:
                        type: integer
                        example: 50
                      failed:
                        type: integer
                        example: 10
        "400":
          description: Invalid campaign ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/calls/bulk:
    post:
      tags:
        - Calls
      summary: Initiate bulk calls
      description: |
        Initiates multiple calls for a specific campaign. This endpoint:
        - Validates the campaign exists and is active
        - Creates call records for each contact
        - Initiates calls via Twilio
        - Returns call IDs for tracking
      operationId: initiateBulkCalls
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - campaign_id
                - contacts
              properties:
                campaign_id:
                  type: string
                  description: MongoDB ObjectID of the campaign
                  example: 507f1f77bcf86cd799439011
                language:
                  type: string
                  description: Override campaign language (optional)
                  example: es
                contacts:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required:
                      - phone_number
                      - name
                    properties:
                      phone_number:
                        type: string
                        description: Phone number in E.164 format
                        example: "+1234567890"
                      name:
                        type: string
                        description: Customer name
                        example: John Doe
            examples:
              single_contact:
                summary: Single contact
                value:
                  campaign_id: "507f1f77bcf86cd799439011"
                  contacts:
                    - phone_number: "+1234567890"
                      name: "John Doe"
              multiple_contacts:
                summary: Multiple contacts with language override
                value:
                  campaign_id: "507f1f77bcf86cd799439011"
                  language: "es"
                  contacts:
                    - phone_number: "+1234567890"
                      name: "John Doe"
                    - phone_number: "+0987654321"
                      name: "Jane Smith"
      responses:
        "200":
          description: Calls initiated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Bulk calls initiated
                  total:
                    type: integer
                    example: 2
                  successful:
                    type: integer
                    example: 2
                  failed:
                    type: integer
                    example: 0
                  call_ids:
                    type: array
                    items:
                      type: string
                      example: 507f1f77bcf86cd799439012
        "400":
          description: Invalid request (campaign not found, inactive campaign, or validation error)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/calls/{id}:
    get:
      tags:
        - Calls
      summary: Get call status
      description: Retrieves the status and logs of a specific call
      operationId: getCallStatus
      parameters:
        - name: id
          in: path
          required: true
          description: Call ID (MongoDB ObjectID)
          schema:
            type: string
            example: 507f1f77bcf86cd799439012
      responses:
        "200":
          description: Call details with logs
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Call"
                  - type: object
                    properties:
                      logs:
                        type: array
                        items:
                          $ref: "#/components/schemas/CallLog"
        "400":
          description: Invalid call ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Call not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Call:
      type: object
      properties:
        _id:
          type: string
          description: MongoDB ObjectID
          example: 507f1f77bcf86cd799439012
        campaign_id:
          type: string
          description: Reference to campaign ObjectID
          example: 507f1f77bcf86cd799439011
        phone_number:
          type: string
          example: "+1234567890"
        customer_name:
          type: string
          example: John Doe
        status:
          type: string
          enum: [pending, initiated, in-progress, completed, failed]
          example: completed
        twilio_call_sid:
          type: string
          description: Twilio's unique call identifier
          example: CA1234567890abcdef1234567890abcdef
        language:
          type: string
          example: en
        duration:
          type: integer
          description: Call duration in seconds
          example: 45
        error_message:
          type: string
          nullable: true
          example: null
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T10:35:00Z"
    CallLog:
      type: object
      properties:
        _id:
          type: string
          description: MongoDB ObjectID
          example: 507f1f77bcf86cd799439013
        call_id:
          type: string
          description: Reference to call ObjectID
          example: 507f1f77bcf86cd799439012
        event:
          type: string
          example: user_input
        details:
          type: string
          example: User selected option 1 - Product Information
        user_input:
          type: string
          nullable: true
          example: "1"
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:32:00Z"
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid request
        message:
          type: string
          example: Campaign ID is required
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (not currently implemented)
# Security is currently not enabled but can be added
# security:
#   - ApiKeyAuth: []